requirements.txt
qrcode==7.4.2
Pillow==10.4.0
cairosvg==2.7.1